+++ RELEASE_NOTES.md
@@ -1,3 +1,8 @@
-### New in 1.0.0-rc1 - (Not yet released)
+### New in 1.0.0-rc2 - (Released 2014-01-19)
+* Removed Process.Start call to allow use with [canopy](http://lefthandedgoat.github.io/canopy/)
+* Updated Script sample to show use with canopy
+* Updated tutorial
+
+### New in 1.0.0-rc1 - (Released 2014-01-18)
 * Release candidate
 * Factored out OWIN-based library from HTML and JavaScript files